refactor: move group config into core
